pyproject.toml

[tool.poetry]
name = "redactor"
version = "0.1.0"
description = "Redacts sensitive information from documents"
authors = ["Allan Holl <allan.holly@somerset.gov.uk>"]
readme = "README.md"
packages = [{ include = "redactor", from = "src" }]

[tool.poetry.dependencies]
python = ">=3.12"
python-dotenv = ">=1.1.1,<2.0.0"
azure-ai-documentintelligence = ">=1.0.2,<2.0.0"
azure-ai-textanalytics = ">=5.3.0,<6.0.0"
pymupdf = ">=1.26.4,<2.0.0"
openai = ">=1.102.0,<2.0.0"
streamlit = "==1.40.0"
pillow = ">=11.3.0,<12.0.0"
fuzzywuzzy = ">=0.18.0,<0.19.0"
levenshtein = ">=0.27.1,<0.28.0"
streamlit-drawable-canvas = ">=0.9.3,<0.10.0"
annoy = "^1.17.3"
uvicorn = "^0.40.0"
fastapi = "^0.128.0"
pydantic = "^2.12.5"
weasyprint = "^67.0"
spire-pdf = "^11.12.1"

[tool.poetry.group.dev.dependencies]
pytest = "^8.0"

[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]